simplfy test func
