+++ metrics/testing.go
@@ -1,0 +1,48 @@
+/*
+Copyright 2019 The Knative Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package metrics
+
+import (
+	"contrib.go.opencensus.io/exporter/stackdriver"
+	"go.opencensus.io/stats/view"
+)
+
+var (
+	testGcpMetadata = gcpMetadata{
+		project:  "test-project",
+		location: "test-location",
+		cluster:  "test-cluster",
+	}
+)
+
+// SetTestingMetricsConfig initialize the package level config with a empty value
+// for unit tests.
+func SetTestingMetricsConfig() {
+	setCurMetricsConfig(&metricsConfig{})
+}
+
+func fakeGcpMetadataFun() *gcpMetadata {
+	return &testGcpMetadata
+}
+
+type fakeExporter struct{}
+
+func (fe *fakeExporter) ExportView(vd *view.Data) {}
+
+func newFakeExporter(o stackdriver.Options) (view.Exporter, error) {
+	return &fakeExporter{}, nil
+}
